package refactoring: solvis.inversion_solution -> solvis.solution;
