--- docs/api/inversion_solution/composite_solution.md
@@ -1,1 +1,0 @@
-::: solvis.inversion_solution.composite_solution
