--- docs/api/inversion_solution/dataframe_models.md
@@ -1,44 +1,0 @@
-::: solvis.inversion_solution.dataframe_models
-    options:
-       inherited_members: false
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.FaultSectionSchemaBase
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.FaultSectionSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.FaultSectionRuptureRateSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.FaultSectionWithSolutionSlipRate
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.ParentFaultParticipationSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.RuptureSectionSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.RuptureSectionsWithRuptureRatesSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.RupturesWithRuptureRatesSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.SectionParticipationSchema
-    options:
-       members: false
-
-::: solvis.inversion_solution.dataframe_models.RuptureRateSchema
-    options:
-       members: false
