--- docs/api/inversion_solution/fault_system_solution_file.md
@@ -1,1 +1,0 @@
-::: solvis.inversion_solution.fault_system_solution_file
