--- docs/api/inversion_solution/inversion_solution.md
@@ -1,10 +1,0 @@
-::: solvis.inversion_solution.inversion_solution
-    options:
-       members: false
-
-::: solvis.inversion_solution.inversion_solution.InversionSolution
-<!--     options:
-        # inherited_members: false
-        members: true
-        members_order: alphabetical -->
-
