--- docs/api/inversion_solution/named_fault.md
@@ -1,1 +1,0 @@
-::: solvis.inversion_solution.named_fault
