--- docs/api/inversion_solution/solution_surfaces_builder.md
@@ -1,1 +1,0 @@
-::: solvis.inversion_solution.solution_surfaces_builder
