--- docs/api/inversion_solution/typing.md
@@ -1,1 +1,0 @@
-::: solvis.inversion_solution.typing
