+++ docs/api/solution/composite_solution.md
@@ -1,0 +1,1 @@
+::: solvis.solution.composite_solution
