+++ docs/api/solution/fault_system_solution/fault_system_solution.md
@@ -1,0 +1,13 @@
+::: solvis.solution.fault_system_solution.fault_system_solution
+<!-- 	options:
+		members: false
+
+::: solvis.solution.fault_system_solution.FaultSystemSolution
+    options:
+       inherited_members: false
+       members_order: alphabetical
+       filters:
+       	- "![A-Z]+"
+       	- "!set_props"
+       	- "!^_"
+       group_by_category: false -->
