+++ docs/api/solution/fault_system_solution/fault_system_solution_file.md
@@ -1,0 +1,1 @@
+::: solvis.solution.fault_system_solution.fault_system_solution_file
