+++ docs/api/solution/fault_system_solution/fault_system_solution_model.md
@@ -1,0 +1,12 @@
+::: solvis.solution.fault_system_solution.fault_system_solution_model
+    options:
+        members: false
+
+::: solvis.solution.fault_system_solution.fault_system_solution_model.FaultSystemSolutionModel
+    options:
+        inherited_members: true
+        filters:
+        - "!^_[^_]"
+        - "!^get_"  # Deprecated function signatures
+        members_order: alphabetical
+        group_by_category: false
