+++ docs/api/solution/fault_system_solution/index.md
@@ -1,0 +1,1 @@
+::: solvis.solution.fault_system_solution
