+++ docs/api/solution/index.md
@@ -1,0 +1,1 @@
+::: solvis.solution
