+++ docs/api/solution/inversion_solution/index.md
@@ -1,0 +1,1 @@
+::: solvis.solution.inversion_solution
