+++ docs/api/solution/inversion_solution/inversion_solution.md
@@ -1,0 +1,3 @@
+::: solvis.solution.inversion_solution.inversion_solution
+
+
