+++ docs/api/solution/inversion_solution/inversion_solution_file.md
@@ -1,4 +1,4 @@
-::: solvis.inversion_solution.inversion_solution_file
+::: solvis.solution.inversion_solution.inversion_solution_file
 <!--     options:
         docstring_options:
             ignore_init_summary: false
