+++ docs/api/solution/named_fault.md
@@ -1,0 +1,1 @@
+::: solvis.solution.named_fault
