+++ docs/api/solution/typing.md
@@ -1,0 +1,1 @@
+::: solvis.solution.typing
