--- solvis/inversion_solution/fault_system_solution/__init__.py
@@ -1,2 +1,0 @@
-from .fault_system_solution import FaultSystemSolution
-from .fault_system_solution_model import FaultSystemSolutionModel
