+++ solvis/solution/solution_surfaces_builder.py
@@ -11,8 +11,8 @@ from solvis.geometry import create_surface, dip_direction, fault_surface_3d
 # from .typing import InversionSolutionProtocol
 
 if TYPE_CHECKING:
-    from solvis.inversion_solution import InversionSolution
     from solvis.fault_system_solution import FaultSystemSolution
+    from solvis.inversion_solution import InversionSolution
 
 log = logging.getLogger(__name__)
