+++ test/filter/test_chainable_set.py
@@ -1,7 +1,7 @@
 import pytest
 
 from solvis.filter.chainable_set_base import ChainableSetBase
-from solvis.inversion_solution.typing import SetOperationEnum
+from solvis.solution.typing import SetOperationEnum
 
 
 class FilterExampleClass(ChainableSetBase):
